en cours

+++ ex02/table.sql
@@ -3,7 +3,9 @@ CREATE TABLE data_2022_oct (
     event_type VARCHAR(4),
     product_id INT,
     price NUMERIC(10, 2),
-    user_id INT,
+    user_id BIGINT,
     user_session UUID
 )
+
+
